Moved clusters/actions to integration-tests/lib
