+++ chain-signatures/contract/Cargo.toml
@@ -11,10 +11,8 @@ borsh = "1.5.3"
 near-sdk = { version = "5.6.0", features = ["legacy", "unit-testing"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-schemars = "0.8"
 k256 = { version = "0.13.1", features = ["sha256", "ecdsa", "serde", "arithmetic", "expose-field"] }
 crypto-shared = { path = "../crypto-shared" }
-near-gas = { version = "0.3.0", features = ["serde", "borsh", "schemars"] }
 thiserror = "1"
 
 [dev-dependencies]
