+++ chain-signatures/node/Cargo.toml
@@ -12,7 +12,6 @@ anyhow = { version = "1", features = ["backtrace"] }
 async-trait = "0.1"
 aws-config = "1.4"
 aws-sdk-s3 = "1.29"
-aws-types = "1.2"
 axum = { version = "0.6.19" }
 axum-extra = "0.7"
 borsh = "1.5.3"
@@ -32,7 +31,6 @@ local-ip-address = "0.5.4"
 rand = "0.8"
 reqwest = { version = "0.11.16", features = ["blocking", "json"] }
 semver = "1.0.23"
-sha2 = "0.10.8"
 sha3 = "0.10.8"
 serde = { version = "1", features = ["derive"] }
 serde_bytes = "0.11.15"
@@ -58,7 +56,6 @@ mpc-contract = { path = "../contract" }
 mpc-keys = { path = "../keys" }
 crypto-shared = { path = "../crypto-shared" }
 
-itertools = "0.12.0"
 http = "1.1.0"
 prometheus = { version = "0.13.3" }
 once_cell = "1.13.1"
