+++ integration-tests/src/actions/mod.rs
@@ -3,6 +3,7 @@ pub mod wait;
 pub mod wait_for;
 
 use crate::cluster::Cluster;
+use crate::containers::LakeIndexer;
 
 use anyhow::Context as _;
 use cait_sith::FullSignature;
@@ -10,14 +11,11 @@ use crypto_shared::ScalarExt;
 use crypto_shared::{derive_epsilon, derive_key};
 use elliptic_curve::sec1::ToEncodedPoint;
 use k256::ecdsa::VerifyingKey;
-use k256::elliptic_curve::ops::{Invert, Reduce};
 use k256::elliptic_curve::point::AffineCoordinates;
 use k256::elliptic_curve::sec1::FromEncodedPoint;
-use k256::elliptic_curve::ProjectivePoint;
 use k256::{AffinePoint, EncodedPoint, Scalar, Secp256k1};
 use mpc_contract::errors::SignError;
 use mpc_contract::primitives::SignRequest;
-use mpc_node::kdf::into_eth_sig;
 use near_crypto::InMemorySigner;
 use near_fetch::ops::AsyncTransactionStatus;
 use near_fetch::ops::Function;
@@ -25,14 +23,10 @@ use near_workspaces::types::Gas;
 use near_workspaces::types::NearToken;
 use near_workspaces::Account;
 use rand::Rng;
-use secp256k1::XOnlyPublicKey;
 use wait_for::{SignatureError, WaitForError};
 
 use std::time::Duration;
 
-const CHAIN_ID_ETH: u64 = 31337;
-
-use integration_tests::containers::LakeIndexer;
 use k256::{
     ecdsa::{Signature as RecoverableSignature, Signature as K256Signature},
     PublicKey as K256PublicKey,
@@ -212,147 +206,8 @@ pub async fn batch_duplicate_signature_production(nodes: &Cluster) -> anyhow::Re
     Ok(())
 }
 
-// This test hardcodes the output of the signing process and checks that everything verifies as expected
-// If you find yourself changing the constants in this test you are likely breaking backwards compatibility
-#[tokio::test]
-async fn signatures_havent_changed() {
-    let big_r = "03f13a99141ce0a4043a7c02afdec6d52f25c6b3de01967acc5cf4a3fa43801589";
-    let s = "39e5631fcc06ffccf8469a3cdcdce0651ebafd998a4280ebbf5dc24a749c98fb";
-    let mpc_key = "04b5695a882aeaf36bf3933e21911b5cbcceae7fd7cb424f3ea221c7e8d390aad4ad2c1a427faec960f22a5442739c0a04fd64ab7ce4c93980417bd3d1d8bc04ea";
-    let account_id = "dev-20240719125040-80075249096169.test.near";
-    let payload_hash: [u8; 32] =
-        hex::decode("49f32740939bfdcbd8d1786075df7aca384381ec203975c3a6c1fd80acddcd4c")
-            .unwrap()
-            .try_into()
-            .unwrap();
-
-    let payload_hash_scalar = Scalar::from_bytes(payload_hash).unwrap();
-
-    // Derive and convert user pk
-    let mpc_pk = hex::decode(mpc_key).unwrap();
-    let mpc_pk = EncodedPoint::from_bytes(mpc_pk).unwrap();
-    let mpc_pk = AffinePoint::from_encoded_point(&mpc_pk).unwrap();
-
-    let account_id = account_id.parse().unwrap();
-    let derivation_epsilon: k256::Scalar = derive_epsilon(&account_id, "test");
-    let user_pk: AffinePoint = derive_key(mpc_pk, derivation_epsilon);
-    let user_pk_y_parity = match user_pk.y_is_odd().unwrap_u8() {
-        0 => secp256k1::Parity::Even,
-        1 => secp256k1::Parity::Odd,
-        _ => unreachable!(),
-    };
-    let user_pk_x = x_coordinate::<k256::Secp256k1>(&user_pk);
-    let user_pk_x: XOnlyPublicKey = XOnlyPublicKey::from_slice(&user_pk_x.to_bytes()).unwrap();
-    let user_secp_pk: secp256k1::PublicKey =
-        secp256k1::PublicKey::from_x_only_public_key(user_pk_x, user_pk_y_parity);
-    let user_address_from_pk = public_key_to_address(&user_secp_pk);
-
-    // Prepare R ans s signature values
-    let big_r = hex::decode(big_r).unwrap();
-    let big_r = EncodedPoint::from_bytes(big_r).unwrap();
-    let big_r = AffinePoint::from_encoded_point(&big_r).unwrap();
-    let big_r_y_parity = big_r.y_is_odd().unwrap_u8() as i32;
-    assert!(big_r_y_parity == 0 || big_r_y_parity == 1);
-
-    let s = hex::decode(s).unwrap().try_into().unwrap();
-    let s = k256::Scalar::from_bytes(s).unwrap();
-    let r = x_coordinate::<k256::Secp256k1>(&big_r);
-
-    let signature = cait_sith::FullSignature::<Secp256k1> { big_r, s };
-
-    let multichain_sig = into_eth_sig(
-        &user_pk,
-        &signature.big_r,
-        &signature.s,
-        payload_hash_scalar,
-    )
-    .unwrap();
-
-    // Check signature using cait-sith tooling
-    let is_signature_valid_for_user_pk = signature.verify(&user_pk, &payload_hash_scalar);
-    let is_signature_valid_for_mpc_pk = signature.verify(&mpc_pk, &payload_hash_scalar);
-    let another_user_pk = derive_key(mpc_pk, derivation_epsilon + k256::Scalar::ONE);
-    let is_signature_valid_for_another_user_pk =
-        signature.verify(&another_user_pk, &payload_hash_scalar);
-    assert!(is_signature_valid_for_user_pk);
-    assert!(!is_signature_valid_for_mpc_pk);
-    assert!(!is_signature_valid_for_another_user_pk);
-
-    // Check signature using ecdsa tooling
-    let k256_sig = k256::ecdsa::Signature::from_scalars(r, s).unwrap();
-    let user_pk_k256: k256::elliptic_curve::PublicKey<Secp256k1> =
-        k256::PublicKey::from_affine(user_pk).unwrap();
-
-    let ecdsa_local_verify_result = verify(
-        &k256::ecdsa::VerifyingKey::from(&user_pk_k256),
-        &payload_hash,
-        &k256_sig,
-    );
-    assert!(ecdsa_local_verify_result.is_ok());
-
-    // TODO: fix
-    // let ecdsa_signature: ecdsa::Signature<Secp256k1> =
-    //     ecdsa::Signature::from_scalars(r, s).unwrap();
-    // let ecdsa_verify_result = ecdsa::signature::Verifier::verify(
-    //     &k256::ecdsa::VerifyingKey::from(&user_pk_k256),
-    //     &payload_hash_reversed,
-    //     &ecdsa_signature,
-    // );
-    // assert!(ecdsa_verify_result.is_ok());
-    // let k256_verify_key = k256::ecdsa::VerifyingKey::from(&user_pk_k256);
-    // let k256_verify_result = k256_verify_key.verify(&payload_hash_reversed, &k256_sig);
-    // assert!(k256_verify_result.is_ok());
-
-    // Check signature using etheres tooling
-    let ethers_r = ethers_core::types::U256::from_big_endian(r.to_bytes().as_slice());
-    let ethers_s = ethers_core::types::U256::from_big_endian(s.to_bytes().as_slice());
-    let ethers_v = to_eip155_v(multichain_sig.recovery_id, CHAIN_ID_ETH);
-
-    let signature = ethers_core::types::Signature {
-        r: ethers_r,
-        s: ethers_s,
-        v: ethers_v,
-    };
-
-    let verifying_user_pk = ecdsa::VerifyingKey::from(&user_pk_k256);
-    let user_address_ethers: ethers_core::types::H160 =
-        ethers_core::utils::public_key_to_address(&verifying_user_pk);
-
-    assert!(signature.verify(payload_hash, user_address_ethers).is_ok());
-
-    // Check if recovered address is the same as the user address
-    let signature_for_recovery: [u8; 64] = {
-        let mut signature = [0u8; 64]; // TODO: is there a better way to get these bytes?
-        signature[..32].copy_from_slice(&r.to_bytes());
-        signature[32..].copy_from_slice(&s.to_bytes());
-        signature
-    };
-
-    let recovered_from_signature_address_web3 = web3::signing::recover(
-        &payload_hash,
-        &signature_for_recovery,
-        multichain_sig.recovery_id as i32,
-    )
-    .unwrap();
-    assert_eq!(user_address_from_pk, recovered_from_signature_address_web3);
-
-    let recovered_from_signature_address_ethers = signature.recover(payload_hash).unwrap();
-    assert_eq!(
-        user_address_from_pk,
-        recovered_from_signature_address_ethers
-    );
-
-    let recovered_from_signature_address_local_function = recover(signature, payload_hash).unwrap();
-    assert_eq!(
-        user_address_from_pk,
-        recovered_from_signature_address_local_function
-    );
-
-    assert_eq!(user_address_from_pk, user_address_ethers);
-}
-
 /// Get the x coordinate of a point, as a scalar
-pub(crate) fn x_coordinate<C: cait_sith::CSCurve>(point: &C::AffinePoint) -> C::Scalar {
+pub fn x_coordinate<C: cait_sith::CSCurve>(point: &C::AffinePoint) -> C::Scalar {
     <C::Scalar as k256::elliptic_curve::ops::Reduce<<C as k256::elliptic_curve::Curve>::Uint>>::reduce_bytes(&point.x())
 }
 
@@ -427,67 +282,223 @@ pub fn public_key_to_address(public_key: &secp256k1::PublicKey) -> web3::types::
     web3::types::Address::from_slice(&hash[12..])
 }
 
-fn verify(
-    key: &VerifyingKey,
-    msg: &[u8],
-    sig: &k256::ecdsa::Signature,
-) -> Result<(), &'static str> {
-    let q = ProjectivePoint::<Secp256k1>::from(key.as_affine());
-    let z = ecdsa::hazmat::bits2field::<Secp256k1>(msg).unwrap();
+#[cfg(test)]
+mod tests {
+    use crypto_shared::{derive_epsilon, derive_key, ScalarExt as _};
+    use elliptic_curve::sec1::FromEncodedPoint as _;
+    use k256::ecdsa::VerifyingKey;
+    use k256::elliptic_curve::ops::{Invert, Reduce};
+    use k256::elliptic_curve::point::AffineCoordinates;
+    use k256::elliptic_curve::ProjectivePoint;
+    use k256::{AffinePoint, EncodedPoint, Scalar};
 
-    // &k256::FieldBytes::from_slice(&k256::Scalar::from_bytes(msg).to_bytes()),
-    verify_prehashed(&q, &z, sig)
-}
+    use super::{public_key_to_address, recover, x_coordinate};
 
-fn verify_prehashed(
-    q: &ProjectivePoint<Secp256k1>,
-    z: &k256::FieldBytes,
-    sig: &k256::ecdsa::Signature,
-) -> Result<(), &'static str> {
-    // let z: Scalar = Scalar::reduce_bytes(z);
-    let z =
+    // This test hardcodes the output of the signing process and checks that everything verifies as expected
+    // If you find yourself changing the constants in this test you are likely breaking backwards compatibility
+    #[test]
+    fn signatures_havent_changed() {
+        const CHAIN_ID_ETH: u64 = 31337;
+
+        let big_r = "03f13a99141ce0a4043a7c02afdec6d52f25c6b3de01967acc5cf4a3fa43801589";
+        let s = "39e5631fcc06ffccf8469a3cdcdce0651ebafd998a4280ebbf5dc24a749c98fb";
+        let mpc_key = "04b5695a882aeaf36bf3933e21911b5cbcceae7fd7cb424f3ea221c7e8d390aad4ad2c1a427faec960f22a5442739c0a04fd64ab7ce4c93980417bd3d1d8bc04ea";
+        let account_id = "dev-20240719125040-80075249096169.test.near";
+        let payload_hash: [u8; 32] =
+            hex::decode("49f32740939bfdcbd8d1786075df7aca384381ec203975c3a6c1fd80acddcd4c")
+                .unwrap()
+                .try_into()
+                .unwrap();
+
+        let payload_hash_scalar = Scalar::from_bytes(payload_hash).unwrap();
+
+        // Derive and convert user pk
+        let mpc_pk = hex::decode(mpc_key).unwrap();
+        let mpc_pk = EncodedPoint::from_bytes(mpc_pk).unwrap();
+        let mpc_pk = AffinePoint::from_encoded_point(&mpc_pk).unwrap();
+
+        let account_id = account_id.parse().unwrap();
+        let derivation_epsilon: k256::Scalar = derive_epsilon(&account_id, "test");
+        let user_pk: AffinePoint = derive_key(mpc_pk, derivation_epsilon);
+        let user_pk_y_parity = match user_pk.y_is_odd().unwrap_u8() {
+            0 => secp256k1::Parity::Even,
+            1 => secp256k1::Parity::Odd,
+            _ => unreachable!(),
+        };
+        let user_pk_x = x_coordinate::<k256::Secp256k1>(&user_pk);
+        let user_pk_x = secp256k1::XOnlyPublicKey::from_slice(&user_pk_x.to_bytes()).unwrap();
+        let user_secp_pk: secp256k1::PublicKey =
+            secp256k1::PublicKey::from_x_only_public_key(user_pk_x, user_pk_y_parity);
+        let user_address_from_pk = public_key_to_address(&user_secp_pk);
+
+        // Prepare R ans s signature values
+        let big_r = hex::decode(big_r).unwrap();
+        let big_r = EncodedPoint::from_bytes(big_r).unwrap();
+        let big_r = AffinePoint::from_encoded_point(&big_r).unwrap();
+        let big_r_y_parity = big_r.y_is_odd().unwrap_u8() as i32;
+        assert!(big_r_y_parity == 0 || big_r_y_parity == 1);
+
+        let s = hex::decode(s).unwrap().try_into().unwrap();
+        let s = k256::Scalar::from_bytes(s).unwrap();
+        let r = x_coordinate::<k256::Secp256k1>(&big_r);
+
+        let signature = cait_sith::FullSignature::<k256::Secp256k1> { big_r, s };
+
+        let multichain_sig = mpc_node::kdf::into_eth_sig(
+            &user_pk,
+            &signature.big_r,
+            &signature.s,
+            payload_hash_scalar,
+        )
+        .unwrap();
+
+        // Check signature using cait-sith tooling
+        let is_signature_valid_for_user_pk = signature.verify(&user_pk, &payload_hash_scalar);
+        let is_signature_valid_for_mpc_pk = signature.verify(&mpc_pk, &payload_hash_scalar);
+        let another_user_pk = derive_key(mpc_pk, derivation_epsilon + k256::Scalar::ONE);
+        let is_signature_valid_for_another_user_pk =
+            signature.verify(&another_user_pk, &payload_hash_scalar);
+        assert!(is_signature_valid_for_user_pk);
+        assert!(!is_signature_valid_for_mpc_pk);
+        assert!(!is_signature_valid_for_another_user_pk);
+
+        // Check signature using ecdsa tooling
+        let k256_sig = k256::ecdsa::Signature::from_scalars(r, s).unwrap();
+        let user_pk_k256: k256::elliptic_curve::PublicKey<k256::Secp256k1> =
+            k256::PublicKey::from_affine(user_pk).unwrap();
+
+        let ecdsa_local_verify_result = verify(
+            &k256::ecdsa::VerifyingKey::from(&user_pk_k256),
+            &payload_hash,
+            &k256_sig,
+        );
+        assert!(ecdsa_local_verify_result.is_ok());
+
+        // TODO: fix
+        // let ecdsa_signature: ecdsa::Signature<Secp256k1> =
+        //     ecdsa::Signature::from_scalars(r, s).unwrap();
+        // let ecdsa_verify_result = ecdsa::signature::Verifier::verify(
+        //     &k256::ecdsa::VerifyingKey::from(&user_pk_k256),
+        //     &payload_hash_reversed,
+        //     &ecdsa_signature,
+        // );
+        // assert!(ecdsa_verify_result.is_ok());
+        // let k256_verify_key = k256::ecdsa::VerifyingKey::from(&user_pk_k256);
+        // let k256_verify_result = k256_verify_key.verify(&payload_hash_reversed, &k256_sig);
+        // assert!(k256_verify_result.is_ok());
+
+        // Check signature using etheres tooling
+        let ethers_r = ethers_core::types::U256::from_big_endian(r.to_bytes().as_slice());
+        let ethers_s = ethers_core::types::U256::from_big_endian(s.to_bytes().as_slice());
+        let ethers_v = to_eip155_v(multichain_sig.recovery_id, CHAIN_ID_ETH);
+
+        let signature = ethers_core::types::Signature {
+            r: ethers_r,
+            s: ethers_s,
+            v: ethers_v,
+        };
+
+        let verifying_user_pk = ecdsa::VerifyingKey::from(&user_pk_k256);
+        let user_address_ethers: ethers_core::types::H160 =
+            ethers_core::utils::public_key_to_address(&verifying_user_pk);
+
+        assert!(signature.verify(payload_hash, user_address_ethers).is_ok());
+
+        // Check if recovered address is the same as the user address
+        let signature_for_recovery: [u8; 64] = {
+            let mut signature = [0u8; 64]; // TODO: is there a better way to get these bytes?
+            signature[..32].copy_from_slice(&r.to_bytes());
+            signature[32..].copy_from_slice(&s.to_bytes());
+            signature
+        };
+
+        let recovered_from_signature_address_web3 = web3::signing::recover(
+            &payload_hash,
+            &signature_for_recovery,
+            multichain_sig.recovery_id as i32,
+        )
+        .unwrap();
+        assert_eq!(user_address_from_pk, recovered_from_signature_address_web3);
+
+        let recovered_from_signature_address_ethers = signature.recover(payload_hash).unwrap();
+        assert_eq!(
+            user_address_from_pk,
+            recovered_from_signature_address_ethers
+        );
+
+        let recovered_from_signature_address_local_function =
+            recover(signature, payload_hash).unwrap();
+        assert_eq!(
+            user_address_from_pk,
+            recovered_from_signature_address_local_function
+        );
+
+        assert_eq!(user_address_from_pk, user_address_ethers);
+    }
+
+    fn verify(
+        key: &VerifyingKey,
+        msg: &[u8],
+        sig: &k256::ecdsa::Signature,
+    ) -> Result<(), &'static str> {
+        let q = ProjectivePoint::<k256::Secp256k1>::from(key.as_affine());
+        let z = ecdsa::hazmat::bits2field::<k256::Secp256k1>(msg).unwrap();
+
+        // &k256::FieldBytes::from_slice(&k256::Scalar::from_bytes(msg).to_bytes()),
+        verify_prehashed(&q, &z, sig)
+    }
+
+    fn verify_prehashed(
+        q: &ProjectivePoint<k256::Secp256k1>,
+        z: &k256::FieldBytes,
+        sig: &k256::ecdsa::Signature,
+    ) -> Result<(), &'static str> {
+        // let z: Scalar = Scalar::reduce_bytes(z);
+        let z =
         <Scalar as Reduce<<k256::Secp256k1 as k256::elliptic_curve::Curve>::Uint>>::reduce_bytes(z);
-    let (r, s) = sig.split_scalars();
-    let s_inv = *s.invert_vartime();
-    let u1 = z * s_inv;
-    let u2 = *r * s_inv;
-    let reproduced = lincomb(&ProjectivePoint::<Secp256k1>::GENERATOR, &u1, q, &u2).to_affine();
-    let x = reproduced.x();
+        let (r, s) = sig.split_scalars();
+        let s_inv = *s.invert_vartime();
+        let u1 = z * s_inv;
+        let u2 = *r * s_inv;
+        let reproduced =
+            lincomb(&ProjectivePoint::<k256::Secp256k1>::GENERATOR, &u1, q, &u2).to_affine();
+        let x = reproduced.x();
 
-    // println!("------------- verify_prehashed[beg] -------------");
-    // println!("z: {z:#?}");
-    // // println!("r: {r:#?}");
-    // // println!("s: {s:#?}");
-    // println!("s_inv {s_inv:#?}");
-    // println!("u1 {u1:#?}");
-    // println!("u2 {u2:#?}");
-    // println!("reproduced {reproduced:#?}");
-    // println!("reproduced_x {x:?}");
-    // println!("------------- verify_prehashed[end] -------------");
+        // println!("------------- verify_prehashed[beg] -------------");
+        // println!("z: {z:#?}");
+        // // println!("r: {r:#?}");
+        // // println!("s: {s:#?}");
+        // println!("s_inv {s_inv:#?}");
+        // println!("u1 {u1:#?}");
+        // println!("u2 {u2:#?}");
+        // println!("reproduced {reproduced:#?}");
+        // println!("reproduced_x {x:?}");
+        // println!("------------- verify_prehashed[end] -------------");
 
-    let reduced =
+        let reduced =
         <Scalar as Reduce<<k256::Secp256k1 as k256::elliptic_curve::Curve>::Uint>>::reduce_bytes(
             &x,
         );
 
-    //println!("reduced {reduced:#?}");
+        //println!("reduced {reduced:#?}");
 
-    if *r == reduced {
-        Ok(())
-    } else {
-        Err("error")
+        if *r == reduced {
+            Ok(())
+        } else {
+            Err("error")
+        }
     }
-}
 
-fn lincomb(
-    x: &ProjectivePoint<Secp256k1>,
-    k: &Scalar,
-    y: &ProjectivePoint<Secp256k1>,
-    l: &Scalar,
-) -> ProjectivePoint<Secp256k1> {
-    (*x * k) + (*y * l)
-}
+    fn lincomb(
+        x: &ProjectivePoint<k256::Secp256k1>,
+        k: &Scalar,
+        y: &ProjectivePoint<k256::Secp256k1>,
+        l: &Scalar,
+    ) -> ProjectivePoint<k256::Secp256k1> {
+        (*x * k) + (*y * l)
+    }
 
-pub fn to_eip155_v(recovery_id: u8, chain_id: u64) -> u64 {
-    (recovery_id as u64) + 35 + chain_id * 2
+    pub fn to_eip155_v(recovery_id: u8, chain_id: u64) -> u64 {
+        (recovery_id as u64) + 35 + chain_id * 2
+    }
 }
