+++ integration-tests/src/actions/wait_for.rs
@@ -8,8 +8,8 @@ use cait_sith::FullSignature;
 use crypto_shared::SignatureResponse;
 use k256::Secp256k1;
 use near_fetch::ops::AsyncTransactionStatus;
-use near_lake_primitives::CryptoHash;
 use near_primitives::errors::ActionErrorKind;
+use near_primitives::hash::CryptoHash;
 use near_primitives::views::ExecutionOutcomeWithIdView;
 use near_primitives::views::ExecutionStatusView;
 use near_primitives::views::FinalExecutionStatus;
